Adopt F4+F5: the wider `k` wire and the draft coherence gate at the shared insert
Adopted from the refused devin seat after verification. That seat left this
work finished but uncommitted when its subscription lapsed; it is reviewed here
against the two dispositions it was written for and verified before landing.

F4 (transports parse `k` before narrowing it): `k` arrives as `Option<i64>` on
both surfaces — the Tauri `WalkForwardRunRequest` and the MCP
`RunWalkForwardArgs` — and `WalkForwardRequest` carries it through as `i64`, so
the checked conversion happens in ONE place, `FoldScheme::rolling_oos`, which
now takes `i64` and refuses outside `2..=12` with `KOutOfRange { k }` naming the
value as sent. Before this, `-1` and `256` died in transport decoding: Tauri
never reached the command, so `typedError` cast a non-`BusError` payload and the
UI rendered a payload with no `code`/`message`, and MCP answered a decode
failure instead of the intended `field_error` on `k`. The refusals now reach
their typed shapes on both sides (BusError `validation` naming `k`; MCP
`field_error` on `k`), and both transports' tests cover `-1` and `256`.

F5 (validate the draft before advancing certification): `WalkForwardRunDraft`
is public with unconstrained fields, so the SHARED insert path — the one funnel
`save_walk_forward_run` and the coach accept both pass through — now runs
`validate_draft` before a single row is written. It refuses a draft whose
claims disagree with each other: fold count other than the scheme's `k`,
indices that do not cover `0..k` in order, windows that do not tile the recorded
span contiguously, a fold run under a different engine fingerprint than the run
records, or a stored verdict that its own folds contradict (`folds_holding`,
`folds_required`, `pass`). Consistency, not re-derivation: re-running the engine
over fold inputs stays the application's job upstream; this boundary proves the
draft's own claims agree, so neither writer can smuggle an incoherent draft past
the certification pointer update.

Verification of the adopted work: `cargo fmt --check` clean,
`cargo clippy --all-targets -- -D warnings` clean, and
`cargo nextest run --no-fail-fast` at 1174 passed / 2 skipped / 1 failed where
the single failure is issue #177, the known `coach_turn_boundary` flake (same
panic text), green on isolated retry.

One correction, found by that verification: the adopted work made
`tests/coach_walk_forward_gate.rs`'s `passing_draft` helper produce a draft whose
recorded verdict contradicts its folds, and
`a_replay_reports_the_childs_current_pointer` overrode that verdict in place to
build its "later failing run" — a run claiming 0 holding folds while both folded
held, which the new gate correctly refuses. The helper is split into
`draft_with_fold_verdict(inputs, span, folds_hold)`, which derives the run
verdict from its folds (`folds_holding`, the scheme's `folds_required`, `pass`)
exactly as the boundary demands, and the failing case now asks for non-holding
folds instead of asserting a verdict they contradict.

+++ src/application/walk_forward.rs
@@ -66,8 +66,11 @@ pub struct WalkForwardRequest {
     /// The counted span's exclusive upper bound; `None` defaults to the
     /// snapshot's last candle's `close_time`.
     pub to_ms: Option<i64>,
-    /// The fold count; `None` defaults to [`K_DEFAULT`]. In `2..=12`.
-    pub k: Option<u8>,
+    /// The fold count; `None` defaults to [`K_DEFAULT`]. In `2..=12` —
+    /// carried as `i64` (wider than the legal `u8`) so an out-of-range wire
+    /// value reaches [`FoldScheme::rolling_oos`]'s typed `KOutOfRange` refusal
+    /// instead of dying in transport decoding.
+    pub k: Option<i64>,
 }
 
 /// The use case's answer — built from the saved rows, never the in-memory
@@ -531,7 +534,7 @@ where
         }
         .into());
     }
-    let scheme = FoldScheme::rolling_oos(request.k.unwrap_or(K_DEFAULT))?;
+    let scheme = FoldScheme::rolling_oos(request.k.unwrap_or(i64::from(K_DEFAULT)))?;
 
     // ONE blocking task for the snapshot loads, the warm-bar probe, the span
     // resolution, and all K fold runs (a13, #201) — the shared unpersisted
